fix sys catalog id and update tests

+++ src/meta/service/src/hummock_service.rs
@@ -244,7 +244,7 @@ impl HummockManagerService for HummockServiceImpl {
         }
 
         // get internal_table_id by metadata_manger
-        if request.table_id < SYS_CATALOG_START_ID {
+        if request.table_id < SYS_CATALOG_START_ID as u32 {
             // We need to make sure to use the correct table_id to filter sst
             let table_id = TableId::new(request.table_id);
             if let Ok(table_fragment) = self
@@ -259,7 +259,7 @@ impl HummockManagerService for HummockServiceImpl {
         assert!(option
             .internal_table_id
             .iter()
-            .all(|table_id| *table_id < SYS_CATALOG_START_ID),);
+            .all(|table_id| *table_id < SYS_CATALOG_START_ID as u32),);
 
         tracing::info!(
             "Try trigger_manual_compaction compaction_group_id {} option {:?}",
